:hammer: Rename lib to common in bin/
