+++ bin/common.mjs
